refactor: move optimizer passes into their own namespace
