+++ src/silo/query_engine/optimizer/column_narrowing_pass.h
@@ -3,7 +3,7 @@
 #include <vector>
 
 #include "silo/query_engine/operators/query_node.h"
-#include "silo/query_engine/pipeline_pass_base.h"
+#include "silo/query_engine/optimizer/pipeline_pass_base.h"
 #include "silo/schema/database_schema.h"
 
 namespace silo::query_engine::operators {
@@ -16,7 +16,7 @@ class OrderByNode;
 class UnionAllNode;
 }  // namespace silo::query_engine::operators
 
-namespace silo::query_engine {
+namespace silo::query_engine::optimizer {
 
 /// Optimization pass that narrows the set of columns produced by each node down to those
 /// actually required by its parent, pruning unneeded columns from scans and collapsing
@@ -45,4 +45,4 @@ class ColumnNarrowingPass : public PipelinePassBase<ColumnNarrowingPass> {
    operators::QueryNodePtr operator()(operators::UnionAllNode& node);
 };
 
-}  // namespace silo::query_engine
+}  // namespace silo::query_engine::optimizer
