+++ src/silo/query_engine/optimizer/column_narrowing_pass.test.cpp
@@ -1,4 +1,4 @@
-#include "silo/query_engine/column_narrowing_pass.h"
+#include "silo/query_engine/optimizer/column_narrowing_pass.h"
 
 #include <memory>
 #include <optional>
@@ -25,6 +25,7 @@
 #include "silo/storage/column/string_column.h"
 #include "silo/storage/table.h"
 
+using silo::query_engine::optimizer::ColumnNarrowingPass;
 using silo::schema::ColumnIdentifier;
 using silo::schema::ColumnType;
 namespace operators = silo::query_engine::operators;
@@ -106,7 +107,7 @@ operators::TableScanNode& leafScan(operators::QueryNode& node) {
 
 TEST(ColumnNarrowingPassScan, keepsOnlyRequiredColumns) {
    auto scan = makeScan({col("a"), col("b"), col("c")});
-   silo::query_engine::ColumnNarrowingPass pass({col("b")});
+   ColumnNarrowingPass pass({col("b")});
    pass(*scan);
 
    EXPECT_THAT(scanSchema(*scan), ::testing::ElementsAre(col("b")));
@@ -114,7 +115,7 @@ TEST(ColumnNarrowingPassScan, keepsOnlyRequiredColumns) {
 
 TEST(ColumnNarrowingPassScan, removesAllWhenNoneRequired) {
    auto scan = makeScan({col("a"), col("b")});
-   silo::query_engine::ColumnNarrowingPass pass({});
+   ColumnNarrowingPass pass({});
    pass(*scan);
 
    EXPECT_THAT(scanSchema(*scan), ::testing::IsEmpty());
@@ -122,7 +123,7 @@ TEST(ColumnNarrowingPassScan, removesAllWhenNoneRequired) {
 
 TEST(ColumnNarrowingPassScan, keepsAllWhenAllRequired) {
    auto scan = makeScan({col("x"), col("y")});
-   silo::query_engine::ColumnNarrowingPass pass({col("x"), col("y")});
+   ColumnNarrowingPass pass({col("x"), col("y")});
    pass(*scan);
 
    EXPECT_THAT(scanSchema(*scan), ::testing::ElementsAre(col("x"), col("y")));
@@ -135,7 +136,7 @@ TEST(ColumnNarrowingPassProject, narrowsScanToProjectedColumns) {
    operators::QueryNodePtr node =
       std::make_unique<operators::ProjectNode>(std::move(scan), std::vector{col("a")});
 
-   silo::query_engine::ColumnNarrowingPass pass({col("a"), col("b"), col("c")});
+   ColumnNarrowingPass pass({col("a"), col("b"), col("c")});
    if (auto new_node = operators::visit(*node, pass)) {
       node = std::move(new_node);
    }
@@ -157,7 +158,7 @@ TEST(ColumnNarrowingPassAggregate, narrowsScanToGroupByColumns) {
       }
    );
 
-   silo::query_engine::ColumnNarrowingPass pass({col("a"), col("b"), col("c")});
+   ColumnNarrowingPass pass({col("a"), col("b"), col("c")});
    pass(*agg);
 
    EXPECT_THAT(scanSchema(leafScan(*agg)), ::testing::ElementsAre(col("b")));
@@ -175,7 +176,7 @@ TEST(ColumnNarrowingPassAggregate, countStarWithNoGroupByKeepsOneColumn) {
       }
    );
 
-   silo::query_engine::ColumnNarrowingPass pass({col("a"), col("b"), col("c")});
+   ColumnNarrowingPass pass({col("a"), col("b"), col("c")});
    pass(*agg);
 
    // COUNT(*) with no group-by needs only one column to drive the row stream.
@@ -193,7 +194,7 @@ TEST(ColumnNarrowingPassOrderBy, appendsSortKeyToRequired) {
    );
 
    // Only "a" is required from above; the pass must add "c" for sorting.
-   silo::query_engine::ColumnNarrowingPass pass({col("a")});
+   ColumnNarrowingPass pass({col("a")});
    pass(*order);
 
    EXPECT_THAT(scanSchema(leafScan(*order)), ::testing::UnorderedElementsAre(col("a"), col("c")));
@@ -207,7 +208,7 @@ TEST(ColumnNarrowingPassProject, collapsesWhenScanNarrowedToProjectFields) {
       std::make_unique<operators::ProjectNode>(std::move(scan), std::vector{col("a"), col("b")});
    auto fetch = std::make_unique<operators::FetchNode>(std::move(proj), std::nullopt, std::nullopt);
 
-   silo::query_engine::ColumnNarrowingPass pass({col("a"), col("b")});
+   ColumnNarrowingPass pass({col("a"), col("b")});
    pass(*fetch);
 
    // The ProjectNode should be removed; FetchNode's child is now directly the TableScanNode.
@@ -226,7 +227,7 @@ TEST(ColumnNarrowingPassProject, stackedProjectsDoNotPropagateSuperfluousColumns
       std::move(inner_proj), std::vector{col("a"), col("b")}
    );
 
-   silo::query_engine::ColumnNarrowingPass pass({col("a")});
+   ColumnNarrowingPass pass({col("a")});
    if (auto new_node = operators::visit(*outer_proj, pass)) {
       outer_proj = std::move(new_node);
    }
@@ -245,7 +246,7 @@ TEST(ColumnNarrowingPassProject, stackedProjectsAreCollapsedAfterNarrowing) {
    auto fetch =
       std::make_unique<operators::FetchNode>(std::move(outer_proj), std::nullopt, std::nullopt);
 
-   silo::query_engine::ColumnNarrowingPass pass({col("a")});
+   ColumnNarrowingPass pass({col("a")});
    pass(*fetch);
 
    EXPECT_NE(dynamic_cast<operators::TableScanNode*>(fetch->child.get()), nullptr);
@@ -269,7 +270,7 @@ TEST(ColumnNarrowingPassZstdDecompress, keepsSequenceColumnInChildWhenRequired) 
    );
 
    // The parent sees "nuc" as STRING (ZstdDecompressNode output schema).
-   silo::query_engine::ColumnNarrowingPass pass({col("nuc")});
+   ColumnNarrowingPass pass({col("nuc")});
    if (auto new_node = operators::visit(*node, pass)) {
       node = std::move(new_node);
    }
@@ -286,7 +287,7 @@ TEST(ColumnNarrowingPassZstdDecompress, doesNotEliminateNodeWhenSequenceColumnRe
       std::move(scan), std::vector{decompressMapping(nuc_col)}
    );
 
-   silo::query_engine::ColumnNarrowingPass pass({col("nuc")});
+   ColumnNarrowingPass pass({col("nuc")});
    if (auto new_node = operators::visit(*node, pass)) {
       node = std::move(new_node);
    }
@@ -306,7 +307,7 @@ TEST(ColumnNarrowingPassZstdDecompress, prunesUnrequiredSequenceColumns) {
    );
 
    // Only "nuc" is required; "aa" and "id" should be dropped from the child scan.
-   silo::query_engine::ColumnNarrowingPass pass({col("nuc")});
+   ColumnNarrowingPass pass({col("nuc")});
    if (auto new_node = operators::visit(*node, pass)) {
       node = std::move(new_node);
    }
@@ -320,7 +321,7 @@ TEST(ColumnNarrowingPassFetch, propagatesRequiredThroughFetch) {
    auto scan = makeScan({col("a"), col("b"), col("c")});
    auto fetch = std::make_unique<operators::FetchNode>(std::move(scan), std::nullopt, std::nullopt);
 
-   silo::query_engine::ColumnNarrowingPass pass({col("b")});
+   ColumnNarrowingPass pass({col("b")});
    pass(*fetch);
 
    EXPECT_THAT(scanSchema(leafScan(*fetch)), ::testing::ElementsAre(col("b")));
@@ -333,7 +334,7 @@ TEST(ColumnNarrowingPassFilter, propagatesRequiredThroughFilter) {
       std::move(scan), std::make_unique<silo::query_engine::expressions::BoolLiteral>(true)
    );
 
-   silo::query_engine::ColumnNarrowingPass pass({col("b")});
+   ColumnNarrowingPass pass({col("b")});
    pass(*filter);
 
    EXPECT_THAT(scanSchema(leafScan(*filter)), ::testing::ElementsAre(col("b")));
@@ -353,7 +354,7 @@ TEST(ColumnNarrowingPassMap, keepsReferencedColumnAndPrunesOthers) {
    );
    auto map = std::make_unique<operators::MapNode>(std::move(scan), std::move(assignments));
 
-   silo::query_engine::ColumnNarrowingPass pass({col("x")});
+   ColumnNarrowingPass pass({col("x")});
    pass(*map);
 
    EXPECT_THAT(scanSchema(leafScan(*map)), ::testing::ElementsAre(col("b")));
@@ -370,7 +371,7 @@ TEST(ColumnNarrowingPassUnionAll, narrowsBothBranchesIndependently) {
    const auto union_all =
       std::make_unique<operators::UnionAllNode>(std::move(left), std::move(right));
 
-   silo::query_engine::ColumnNarrowingPass pass({col("a")});
+   ColumnNarrowingPass pass({col("a")});
    pass(*union_all);
 
    EXPECT_THAT(scanSchema(leafScan(*union_all->left)), ::testing::ElementsAre(col("a")));
